requirements.txt
Flask==3.0.3
PyMySQL==1.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
